requirements.txt
apache-airflow[amazon]==2.9.3
apache-airflow-providers-amazon==8.24.0
boto3==1.34.144
requests==2.32.3
pandas==2.2.2
pyarrow==16.1.0
